Revert the logic to it's original intention
originally the video was only supposed to play if you await for the attract time for enough time, but for testing reasons, we also made the video play when you press start on the title screen, oh and also, alpha in the player anims was added to Scarlet Engine, so go cuckoo-crazy with it! Wahoo!